refactor: 추가 비동기 LogRepository 구현을 위한 인터페이스 분리
- 기존 LogRepository를 SynchronousLogRepository로 변경 후 인터페이스 구현하도록
- 테스트코드 클래스명 수정
- 새로운 비동기 LogRepository를 AsyncLogRepository 클래스 추가
